minor updates, added URL checking for loading in a URL, added comments for main page on what's left to do (marked with TODO's)

+++ TheatreProjectWeb/src/theatreProject/server/PersistenceImpl.java
@@ -58,7 +58,7 @@ public class PersistenceImpl extends RemoteServiceServlet implements Persistence
 	
 
 	@Override
-	public InventoryObject getInventoryObject(Long ID) {
+	public InventoryObject getInventoryObject(int ID) {
 		PersistenceManager persistenceManager = pmf.getPersistenceManager();
 		try{
 			return persistenceManager.getObjectById(InventoryObject.class, ID);
